net: ieee802154: ca8210: Refuse most of the scan operations
The Cascada 8210 hardware transceiver is kind of a hardMAC which
interfaces with the softMAC and in practice does not support sending
anything else than dataframes. This means we cannot send any BEACON_REQ
during active scans nor any BEACON in general. Refuse these operations
officially so that the user is aware of the limitation.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>