crypto: asymmetric_keys: remove redundant pointer secs
The pointer secs is assigned a value but it is never read. The
pointer is redundant and can be removed. Cleans up clang scan
warning:

crypto/asymmetric_keys/verify_pefile.c:113:14: warning: Although
the value stored to 'secs' is used in the enclosing expression,
the value is never actually read from 'secs' [deadcode.DeadStores]

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>