iwlwifi: mvm: check if SAR GEO is supported before sending command
Older hardware, for instance 3160, do not support SAR GEO offsets.  We
used to check for support before sending the command, but when moving
the command to the init phase, we lost the check.  This causes a
failure when initializing HW that do not support this command.

Fix that by adding a check before sending the command.  Additionally,
fix the caller so that it checks for the return value of the
iwl_mvm_sar_geo_init() function, which it was ignoring.

Fixes: db700bc35703 ("iwlwifi: mvm: check if SAR GEO is supported before sending command")
Cc: stable@vger.kernel.org # 5.15+
Reported-by: Len Brown <lenb@kernel.org>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>