ARM: dts: Add support for Samsung Chagallwifi
Chagallwifi, with product name Samsung Galaxy Tab S 10.5", is based on
Exynos 5420. This device is one of several tablet models released in
2013 - 2014 based on Exynos 5420.

The device tree added here contains support for:

- UART
- eMMC
- SD card
- USB

CCI has been disabled in the hardware, enabling it would require
(de-)soldering a resistor on the board.  Trying to boot with it
enabled in kernel makes the device hang when CCI is probed.
Exynos5420-arndale-octa also has had CCI disabled due to issues [1].

To successfully boot the mainline kernel with the stock bootloader
(SBOOT), an additional hack is needed [2]. The same hack is also
needed to boot exynos4412-i9300 with stock bootloader, and probably
other Samsung devices of similar age.

[1] https://marc.info/?l=linux-arm-kernel&m=141718639200624
[2] https://lore.kernel.org/all/1355276466-18295-1-git-send-email-arve@android.com/

Signed-off-by: Henrik Grimler <henrik@grimler.se>