perf: arm_spe: make the PMSCR CX bit[3] consistent across the session
The ARM SPE driver will seenablet the CX bit in the PMSCR register if it
detects that the profiler (perf tool) has enough capabilities during the
initialization of the PMU event in the "arm_spe_pmu_event_init" callback:

  reg = arm_spe_event_to_pmscr(event);
  if (!perfmon_capable() &&
      (reg & (BIT(SYS_PMSCR_EL1_PA_SHIFT) |
              BIT(SYS_PMSCR_EL1_CX_SHIFT) |
              BIT(SYS_PMSCR_EL1_PCT_SHIFT))))
          return -EACCES;

This implies that the CX bit should remain consistent across the entire
duration of the session. However, the value of the bit is later being
changed (from 0 to 1 and vice versa) in the "arm_spe_pmu_start" callback

Consider this example:

  1. Run a process in the background with privileges in CPU0

    $ taskset --cpu-list 0 sudo dd if=/dev/random of=/dev/null &
    [3] 3806

  2. Begin a perf session without privileges (we shouldn't see CONTEXT packets)

    $ perf record -e arm_spe_0// -C0 -- sleep 1
    $ perf report -D | grep CONTEXT
    .  0000000e:  65 df 0e 00 00                                  CONTEXT 0xedf el2
    .  0000004e:  65 df 0e 00 00                                  CONTEXT 0xedf el2
    .  0000008e:  65 df 0e 00 00                                  CONTEXT 0xedf el2
    [...]

One way to fix this is by caching the value of the CX bit during the
initialization of the PMU event, so that it remains consistent during
the session.

Signed-off-by: German Gomez <german.gomez@arm.com>