net: Fix compiler warnings on 4.19 PREEMPT_RT with xmit_lock_owner
GCC 11 has started complaining about the PREEMPT_RT changes to 2fd949365fe6628fb2
which change the xmit_lock_owner from an integer to a pointer to struct
task_struct. These changes are from the patch:

	     net: move xmit_recursion to per-task variable on -RT

The below patch is ugly but fixes the compiler warnings.

I'd be happy to entertain a better/cleaner solution

Signed-off-by: Clark Williams <williams@redhat.com>