drm/arm/mali: Check for NULL pointer after calling kzalloc
As the possible failure of the kzalloc(), mw_state could be NULL
pointer.
Therefore, it should be better to check it in order to avoid the
dereference of the NULL pointer, like the kzalloc() in
malidp_mw_connector_duplicate_state().
If fails, we can directly use the 'NULL' instead of the
'&mw_state->base' and __drm_atomic_helper_connector_reset() will deal
with it correctly.

Fixes: 8cbc5caf36ef ("drm: mali-dp: Add writeback connector")
Signed-off-by: Jiasheng Jiang <jiasheng@iscas.ac.cn>